# Remove @linearidx in permutedims kernel